+++ router-processor-cdi/src/main/java/org/jboss/elemento/router/processor/CdiCodeGenerator.java
@@ -15,15 +15,8 @@
  */
 package org.jboss.elemento.router.processor;
 
-import java.io.IOException;
-import java.util.List;
-
-import javax.annotation.processing.Filer;
-
 import com.squareup.javapoet.ClassName;
-import com.squareup.javapoet.JavaFile;
 import com.squareup.javapoet.MethodSpec;
-import com.squareup.javapoet.TypeSpec;
 
 import static javax.lang.model.element.Modifier.PUBLIC;
 
@@ -31,30 +24,18 @@ class CdiCodeGenerator extends CodeGenerator {
 
     private static final String BEAN_MANAGER = "beanManager";
 
-    void generateCode(Filer filer, String packageName, List<RouteInfo> routes) throws IOException {
+    @Override
+    MethodSpec.Builder buildConstructor() {
         ClassName beanManagerClass = ClassName.get("org.kie.j2cl.tools.di.core", "BeanManager");
-        ClassName placeClass = ClassName.get("org.jboss.elemento.router", "Place");
-        ClassName placesClass = ClassName.get("org.jboss.elemento.router", "Places");
-
-        MethodSpec.Builder constructorBuilder = MethodSpec.constructorBuilder()
+        return MethodSpec.constructorBuilder()
                 .addModifiers(PUBLIC)
                 .addParameter(beanManagerClass, BEAN_MANAGER)
                 .addStatement("super()");
+    }
 
-        for (RouteInfo route : routes) {
-            String placeName = placeStatement(constructorBuilder, placeClass, route);
-            constructorBuilder.addStatement("add($N, () -> $N.lookupBean($L.class).getInstance())",
-                    placeName, BEAN_MANAGER, route.pageClass);
-        }
-
-        TypeSpec routesImplType = TypeSpec.classBuilder(Names.CLASS)
-                .superclass(placesClass)
-                .addModifiers(PUBLIC)
-                .addMethod(constructorBuilder.build())
-                .build();
-
-        JavaFile javaFile = JavaFile.builder(packageName, routesImplType)
-                .build();
-        javaFile.writeTo(filer);
+    @Override
+    void addPlace(MethodSpec.Builder constructor, String placeName, RouteInfo route) {
+        constructor.addStatement("add($N, () -> $N.lookupBean($L.class).getInstance())",
+                placeName, BEAN_MANAGER, route.pageClass);
     }
 }