+++ router-processor-common/src/main/java/org/jboss/elemento/router/processor/CodeGenerator.java
@@ -19,18 +19,42 @@ import java.io.IOException;
 import java.util.List;
 
 import javax.annotation.processing.Filer;
-
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
+import com.squareup.javapoet.JavaFile;
 import com.squareup.javapoet.MethodSpec;
+import com.squareup.javapoet.TypeSpec;
+
+import static javax.lang.model.element.Modifier.PUBLIC;
 
 abstract class CodeGenerator {
 
     private int counter = 0;
 
-    abstract void generateCode(Filer filer, String packageName, List<RouteInfo> routes) throws IOException;
+    final void generateCode(Filer filer, String packageName, String className, List<RouteInfo> routes) throws IOException {
+        ClassName placeClass = ClassName.get("org.jboss.elemento.router", "Place");
+        ClassName placesClass = ClassName.get("org.jboss.elemento.router", "Places");
 
-    String placeStatement(MethodSpec.Builder constructorBuilder, ClassName placeClass, RouteInfo route) {
+        MethodSpec.Builder constructor = buildConstructor();
+        for (RouteInfo route : routes) {
+            String placeName = createPlace(constructor, placeClass, route);
+            addPlace(constructor, placeName, route);
+        }
+
+        TypeSpec implType = TypeSpec.classBuilder(className)
+                .superclass(placesClass)
+                .addModifiers(PUBLIC)
+                .addMethod(constructor.build())
+                .build();
+
+        JavaFile javaFile = JavaFile.builder(packageName, implType)
+                .build();
+        javaFile.writeTo(filer);
+    }
+
+    abstract MethodSpec.Builder buildConstructor();
+
+    String createPlace(MethodSpec.Builder constructor, ClassName placeClass, RouteInfo route) {
         String placeName = "place" + counter;
 
         CodeBlock.Builder builder = CodeBlock.builder();
@@ -41,9 +65,11 @@ abstract class CodeGenerator {
         if (route.selector != null) {
             builder.add(".root($S)", route.title);
         }
-        constructorBuilder.addStatement(builder.build());
+        constructor.addStatement(builder.build());
 
         counter++;
         return placeName;
     }
+
+    abstract void addPlace(MethodSpec.Builder constructor, String placeName, RouteInfo route);
 }